feature add: show total credit per day when doing comparisons
Credit per work unit is assume to be 1.0.  CPD is calculated by dividing by the total elapsed time of each app for each system and multiplying by the number of seconds.   It is normalied to 1 to you must estimate the average credit for a work unit and multiply the CPD to get an approximate CPD.