Account for unknown GPUs
A number of GPUs are not identified in the raw data.  It appears the problem occurs only on the fastest graphic processors.  For example, the AMD VII appeared to be the only board out of S9150, S9100, S9000 that triggered the problem.  The problem seems to be a function of the number of concurrent work units.  To make the user aware of the problem, the statistics for the unknown work units is not being displayed.