Had to handle GPUs that were unidentified
Fast GPUs such as AMD VII that can concurrently process multiple work units are occasionally missing the GPU identifier. For 4 concurrent work units, when this happens it is normally the last 1 or 2 of the 4 units.  This is mitigated by using the previous GPU# for the missing GPU.
A calculation of average elapsed time was mistakenly used for throughput computation.  That is incorrect as the work units were concurrent in time not sequential.