Allow unknown GPUs to be assigned to a known GPU
The advanced filter option was improved to allow one to select either the previous GPU or to select any of the known GPUs.  By unchecking the assignment one can see the statistics for all the unknown GPUs.  From observation the unknown GPUs are more often than not belonging to the fastest GPU then that GPU can be assigned to pick up all the unknown gpu elapsed times.